// src/lib/enums/product.enum.ts
export enum ProductSize {
  US_4 = "US_4",
  US_4_5 = "US_4.5",
  US_5 = "US_5",
  US_5_5 = "US_5.5",
  US_6 = "US_6",
  US_6_5 = "US_6.5",
  US_7 = "US_7",
  US_7_5 = "US_7.5",
  US_8 = "US_8",
  US_8_5 = "US_8.5",
  US_9 = "US_9",
  US_9_5 = "US_9.5",
  US_10 = "US_10",
}

export enum ProductWeightGram {
  G_1 = 1,
  G_2 = 2,
  G_3 = 3,
  G_4 = 4,
  G_5 = 5,
  G_6 = 6,
  G_8 = 8,
  G_10 = 10,
  G_12 = 12,
  G_15 = 15,
  G_20 = 20,
  G_30 = 30,
  G_50 = 50,
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  RING = "RING",
  NECKLACE = "NECKLACE",
  BRACELET = "BRACELET",
  EARRING = "EARRING",
  OTHER = "OTHER",
}